Renamed StellarProcessorUtils to StellarExecutorUtils
