+++ metron-platform/metron-management/src/test/java/org/apache/metron/management/GrokFunctionsTest.java
@@ -24,7 +24,7 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import java.util.HashMap;
-import static org.apache.metron.common.utils.StellarProcessorUtils.run;
+import static org.apache.metron.common.utils.StellarExecutorUtils.run;
 
 public class GrokFunctionsTest {
   private String grokExpr = "%{NUMBER:timestamp}[^0-9]*%{INT:elapsed} %{IP:ip_src_addr} %{WORD:action}/%{NUMBER:code} %{NUMBER:bytes} %{WORD:method} %{NOTSPACE:url}[^0-9]*(%{IP:ip_dst_addr})?";
