+++ metron-platform/metron-management/src/test/java/org/apache/metron/management/ShellFunctionsTest.java
@@ -24,12 +24,10 @@ import org.apache.metron.common.stellar.shell.StellarExecutor;
 import org.junit.Assert;
 import org.junit.Test;
 
-import java.util.Collections;
 import java.util.HashMap;
-import java.util.List;
 import java.util.Map;
 
-import static org.apache.metron.common.utils.StellarProcessorUtils.run;
+import static org.apache.metron.common.utils.StellarExecutorUtils.run;
 
 public class ShellFunctionsTest {
